Updated unlock to allow for mutiple unlocks and unlock nesting (#37)
Previously, the unlock function did not support multiple unlocks with different configs within a program. Additionally, you could not nest unlock decorators if you wanted to combine multiple configs into one config object. With this change, the unlock function can now do both.

If a user has two functions, each with a different @unlock call, and each pointing to a different config, they can do that. Each Config object created will only have the keys from the YAML file specified in the unlock. These keys can be overwritten from the command line as usual. However, this means that a user can now specify arbitrary keys from the command line without causing an error. These keys go unparsed, and if a user tries to access these unparsed keys that do not exist in a YAML file, they will get an attribute error indicating that they did specify this key from the command line, but they need it to be in the YAML config. Because I am storing the unparsed keys in the config object as a private variable, I created a "to_dict" method for the Config class that controls how the config object is converted to a dictionary, which means that this private attribute is excluded in the conversion. This "to_dict" method was used in place of using the 'vars' function within the instantiate function family so that they do not attempt to add private attributes from the config class into the object the function is attempting to instantiate. Additionally, if a user tries to specify a config from the command line when there are multiple unlocks, they are given a warning that tells them that all the configs have been overwritten with this config. In the future, this can be changed to potentially only overwriting the config from the first call of unlock or some other behavior.

Additionally, I changed the unlock function in a way that allows nesting. This means that a user can add multiple unlock decorators to a single function. The keys from the last decorator applied will overwrite any matching keys from the previous unlock.

Future additions may include allowing a user to specify a prefix in an unlock, meaning that they can nest unlocks with the same key names without having them overwritten. Additionally, with the prefix, it would potentially allow for separate unlocks to be changed via the command line (i.e., the prefix is an unlock identifier). Also, the unlock can be changed so that the user can specify skeleton key keywords like "keyring."

Overall, I think this expands the use cases for skeleton key, allowing for more complex behavior.